Moving the SearchParams type
