+++ src/components/form/LinkToPotentialNode.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 import { Link } from 'react-router-dom';
 import type { LinkProps } from 'react-router-dom';
 
-import { SearchParams } from 'src/hooks/useNavigatePage';
+import { SearchParams } from 'src/features/routing/AppRoutingContext';
 import { Hidden, useNode } from 'src/utils/layout/NodesContext';
 
 type Props = LinkProps & { children?: React.ReactNode };
